Do not clear capture groups that are matched out-of-order
In the regular expression /(?=a(b))(a)b/, it will always be the case
that the first capture group matches a string which *follows* the string
matched by the second capture group. Therefore, it would not be sound
to clear the captured result of the second capture group.

This change was introduced to implement the JS behavior of clearing
capture groups when (re-)entering a quantifier. However, it is not
sound in the presence of look-ahead and TRegex already manages this
clearing by itself. A future commit will implement this behavior
for JONI as well.