Fix re-entering of looping groups in NFATraversalRegexASTVisitor
This fixes regular expressions such as /([a]*?)*/, where the NFA traversal
of the regex AST required us to enter the same looping group multiple times.
This meant abandoning the use of the Group implementation of RegexASTVisitorIterable
in NFATraversalRegexASTVisitor and a finer management of the AST nodes
tracked as visited by the NFATraversalRegexASTVisitor.