Recover from unsupported regular expressions by throwing a JS Error
Since our regular expressions are compiled lazily, at execution time, we cannot
know in advance if our RegExp compilation toolchain will be able to handle a
given regular expression or not. This can lead to bailouts in all of our RegExp
engines during a RegExp's execution, which we should be able to recover from
without crashing.