Joni: Only perform null checks in optional iterations of REPEAT loops
The general fallback strategy for implementing quantifiers (RepeatMatcher in the
ECMAScript spec) is using the REPEAT* bytecode instructions. Currently, null
checks (which verify that we match more than an empty string) were performed on
every iteration: both the mandatory ones and the optional ones (in /a{n,m}/, the
first n iterations are mandatory, the rest is optional). However, null checks
should only be used in optional iterations. This commit modifies the behavior of
the REPEAT* instruction so that they manage null checks themselves based on the
number of the current iteration.

Together with the last commit, this fixed
/[a-z]+(?:(.*)(.*)(.*)(.*)(.*)(.*)(.*)(.*)(.*)(.*)(.*))+/.exec("foo").