Enable patching of the scripting option
Some code to enable the patching of this option was already present in the
codebase. This commit extends the support to both the enabling and disabling
of the scripting builtins and also adds the scripting option among the set
of options which are allowed to vary between the pre-initialized context
and the runtime context.