Use JSRuntime.isNegativeZero() to check for -0.0
In `PowNode::pow()` there are optimizations for specific input values which map `pow()` to the cheaper `sqrt()` function. But for some double inputs like `-0.0` this mapping is not valid and has to be excluded.

However, we can not simply compare a double value `d` with `-0.0` because the JLS (and the IEEE 754 standard) require that positive and negative zeros are considered equal and therefor `0.0 == -0.0`.

In order to really check for a negative zero we have to use `JSRuntime.isNegativeZero()`.